{feature} support angular 13
Rewrite the entire project in javascript, because we're 2022
and there's still no good way to run typescript code when
using ESM and Angular 13 CDK only exposes ESM code.

Jest and ESM is a whole other story, and the jest builder was
a lot more complex to write and maintain, so drop jest support.