{feature} add isOpen variant to screen types
This makes it easier to write tests that span multiple pages,
for example

```ts
class ScreenOne extends ComponentHarness {
	static path = '/one';

	#next = this.locatorFor('#next');

	async goToNext() {
		await (await this.next()).click();
	}
}

class ScreenTwo extends ComponentHarness {
	static async isOpen(page) {
	  return page.url().endsWith('/two');
	}
}

test('navigation', ({open}) => {
  const screenOne = await open(ScreenOne);

  await screenOne.goToNext();

  const screenTwo = await open(ScreenTwo);

  // ...
});
```

Screens with a path or open function can also define an
isOpen function. If that function is present and it yields
a truthy result, then the open function will not be called
or no navigation will be triggered to the path.